requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
absl-py==2.3.1
    # via
    #   keras
    #   tensorboard
    #   tensorflow
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   openai
astunparse==1.6.3
    # via tensorflow
async-timeout==4.0.3
    # via langchain
bibtexparser==1.4.3
    # via pymatgen
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
contourpy==1.3.2
    # via matplotlib
cycler==0.12.1
    # via matplotlib
distro==1.9.0
    # via openai
exceptiongroup==1.3.0
    # via anyio
flatbuffers==25.2.10
    # via tensorflow
fonttools==4.59.2
    # via matplotlib
gast==0.6.0
    # via tensorflow
google-pasta==0.2.0
    # via tensorflow
greenlet==3.2.4
    # via sqlalchemy
grpcio==1.74.0
    # via
    #   tensorboard
    #   tensorflow
h11==0.16.0
    # via httpcore
h5py==3.14.0
    # via
    #   keras
    #   tensorflow
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   langgraph-sdk
    #   langsmith
    #   openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-resources==6.5.2
    # via molsimplify
jiter==0.10.0
    # via openai
joblib==1.5.2
    # via
    #   pymatgen
    #   scikit-learn
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
keras==3.11.3
    # via
    #   molsimplify
    #   tensorflow
kiwisolver==1.4.9
    # via matplotlib
langchain==0.3.27
    # via chemunityqa (pyproject.toml)
langchain-core==0.3.76
    # via
    #   langchain
    #   langchain-openai
    #   langchain-text-splitters
    #   langgraph
    #   langgraph-checkpoint
    #   langgraph-prebuilt
langchain-openai==0.3.33
    # via langchain
langchain-text-splitters==0.3.11
    # via langchain
langgraph==0.6.7
    # via chemunityqa (pyproject.toml)
langgraph-checkpoint==2.1.1
    # via
    #   langgraph
    #   langgraph-prebuilt
langgraph-prebuilt==0.6.4
    # via langgraph
langgraph-sdk==0.2.6
    # via langgraph
langsmith==0.4.27
    # via
    #   langchain
    #   langchain-core
libclang==18.1.1
    # via tensorflow
loguru==0.7.3
    # via chemunityqa (pyproject.toml)
markdown==3.9
    # via tensorboard
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.2
    # via werkzeug
matplotlib==3.10.6
    # via pymatgen
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.5.3
    # via
    #   keras
    #   tensorflow
molsimplify==1.8.0
    # via chemunityqa (pyproject.toml)
monty==2025.3.3
    # via pymatgen
mpmath==1.3.0
    # via sympy
namex==0.1.0
    # via keras
narwhals==2.5.0
    # via plotly
networkx==3.4.2
    # via
    #   molsimplify
    #   pymatgen
numpy==2.2.6
    # via
    #   chemunityqa (pyproject.toml)
    #   contourpy
    #   h5py
    #   keras
    #   matplotlib
    #   ml-dtypes
    #   molsimplify
    #   monty
    #   pandas
    #   pymatgen
    #   scikit-learn
    #   scipy
    #   spglib
    #   tensorboard
    #   tensorflow
    #   xgboost
nvidia-nccl-cu12==2.28.3
    # via xgboost
openai==1.107.1
    # via langchain-openai
openbabel-wheel==3.1.1.22
    # via molsimplify
opt-einsum==3.4.0
    # via tensorflow
optree==0.17.0
    # via keras
orjson==3.11.3
    # via
    #   langgraph-sdk
    #   langsmith
    #   pymatgen
ormsgpack==1.10.0
    # via langgraph-checkpoint
packaging==25.0
    # via
    #   keras
    #   langchain-core
    #   langsmith
    #   matplotlib
    #   plotly
    #   tensorboard
    #   tensorflow
palettable==3.3.3
    # via pymatgen
pandas==2.3.2
    # via
    #   chemunityqa (pyproject.toml)
    #   molsimplify
    #   pymatgen
pillow==11.3.0
    # via
    #   matplotlib
    #   tensorboard
plotly==6.3.0
    # via pymatgen
protobuf==6.32.1
    # via
    #   tensorboard
    #   tensorflow
pydantic==2.11.7
    # via
    #   langchain
    #   langchain-core
    #   langgraph
    #   langsmith
    #   openai
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
pymatgen==2025.6.14
    # via chemunityqa (pyproject.toml)
pyparsing==3.2.3
    # via
    #   bibtexparser
    #   matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
    #   molsimplify
regex==2025.9.1
    # via tiktoken
requests==2.32.5
    # via
    #   langchain
    #   langsmith
    #   pymatgen
    #   requests-toolbelt
    #   tensorflow
    #   tiktoken
requests-toolbelt==1.0.0
    # via langsmith
rich==14.1.0
    # via keras
ruamel-yaml==0.18.15
    # via
    #   monty
    #   pymatgen
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
scikit-learn==1.7.2
    # via
    #   chemunityqa (pyproject.toml)
    #   molsimplify
scipy==1.15.3
    # via
    #   chemunityqa (pyproject.toml)
    #   molsimplify
    #   pymatgen
    #   scikit-learn
    #   xgboost
setuptools==80.9.0
    # via
    #   tensorboard
    #   tensorflow
six==1.17.0
    # via
    #   astunparse
    #   google-pasta
    #   python-dateutil
    #   tensorflow
sniffio==1.3.1
    # via
    #   anyio
    #   openai
spglib==2.6.0
    # via pymatgen
sqlalchemy==2.0.43
    # via langchain
sympy==1.14.0
    # via pymatgen
tabulate==0.9.0
    # via pymatgen
tenacity==9.1.2
    # via langchain-core
tensorboard==2.20.0
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.20.0
    # via molsimplify
termcolor==3.1.0
    # via tensorflow
threadpoolctl==3.6.0
    # via scikit-learn
tiktoken==0.11.0
    # via langchain-openai
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   anyio
    #   exceptiongroup
    #   langchain-core
    #   openai
    #   optree
    #   pydantic
    #   pydantic-core
    #   spglib
    #   sqlalchemy
    #   tensorflow
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
uncertainties==3.2.3
    # via pymatgen
urllib3==2.5.0
    # via requests
werkzeug==3.1.3
    # via tensorboard
wheel==0.45.1
    # via astunparse
wrapt==1.17.3
    # via tensorflow
xgboost==3.0.5
    # via chemunityqa (pyproject.toml)
xxhash==3.5.0
    # via langgraph
zstandard==0.24.0
    # via langsmith
langchain
langchain-openai
langgraph
neo4j
pydantic
pyyaml
